requirements.txt
streamlit
google-cloud-storage
google-cloud-firestore
python-dotenv
python-magic
tenacity
vertexai
google-cloud-aiplatform
libmagic